requirements.txt
pdfplumber
neo4j
python-dotenv
transformers
torch
accelerate
sentencepiece